+++ timeline/src/main/java/com/webapp/timeline/sns/web/PostFileUploadController.java
@@ -1,10 +1,8 @@
 package com.webapp.timeline.sns.web;
 
 
-import com.webapp.timeline.sns.service.PostService;
+import com.webapp.timeline.sns.service.interfaces.PostService;
 import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiImplicitParam;
-import io.swagger.annotations.ApiImplicitParams;
 import io.swagger.annotations.ApiOperation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -17,10 +15,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import springfox.documentation.annotations.ApiIgnore;
 
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.util.Arrays;
 import java.util.Collections;
